Extract evaluation and search to separate files
