+++ CrusherChess/src/evaluation.h
@@ -1,0 +1,278 @@
+#pragma once
+
+
+
+// material scrore
+
+/*
+	♙ =   100   = ♙
+	♘ =   300   = ♙ * 3
+	♗ =   350   = ♙ * 3 + ♙ * 0.5
+	♖ =   500   = ♙ * 5
+	♕ =   1000  = ♙ * 10
+	♔ =   10000 = ♙ * 100
+
+*/
+
+
+
+#include "typedefs.h"
+#include "types.h"
+// material score [game phase][piece]
+constexpr int material_score[2][12] =
+{
+	// opening material score
+	82, 337, 365, 477, 1025, 12000, -82, -337, -365, -477, -1025, -12000,
+
+	// endgame material score
+	94, 281, 297, 512,  936, 12000, -94, -281, -297, -512,  -936, -12000
+};
+
+// game phase scores
+constexpr int opening_phase_score = 6192;
+constexpr int endgame_phase_score = 518;
+
+// game phases
+enum { OPENING, ENDGAME, MIDDELGAME };
+
+
+
+// positional piece scores [game phase][piece][square]
+constexpr int positional_score[2][6][SQUARE_NB] =
+
+// opening positional piece scores //
+{
+	//pawn
+	0,   0,   0,   0,   0,   0,  0,   0,
+	98, 134,  61,  95,  68, 126, 34, -11,
+	-6,   7,  26,  31,  65,  56, 25, -20,
+	-14,  13,   6,  21,  23,  12, 17, -23,
+	-27,  -2,  -5,  12,  17,   6, 10, -25,
+	-26,  -4,  -4, -10,   3,   3, 33, -12,
+	-35,  -1, -20, -23, -15,  24, 38, -22,
+	0,   0,   0,   0,   0,   0,  0,   0,
+
+	// knight
+	-167, -89, -34, -49,  61, -97, -15, -107,
+	-73, -41,  72,  36,  23,  62,   7,  -17,
+	-47,  60,  37,  65,  84, 129,  73,   44,
+	-9,  17,  19,  53,  37,  69,  18,   22,
+	-13,   4,  16,  13,  28,  19,  21,   -8,
+	-23,  -9,  12,  10,  19,  17,  25,  -16,
+	-29, -53, -12,  -3,  -1,  18, -14,  -19,
+	-105, -21, -58, -33, -17, -28, -19,  -23,
+
+	// bishop
+	-29,   4, -82, -37, -25, -42,   7,  -8,
+	-26,  16, -18, -13,  30,  59,  18, -47,
+	-16,  37,  43,  40,  35,  50,  37,  -2,
+	-4,   5,  19,  50,  37,  37,   7,  -2,
+	-6,  13,  13,  26,  34,  12,  10,   4,
+	0,  15,  15,  15,  14,  27,  18,  10,
+	4,  15,  16,   0,   7,  21,  33,   1,
+	-33,  -3, -14, -21, -13, -12, -39, -21,
+
+	// rook
+	32,  42,  32,  51, 63,  9,  31,  43,
+	27,  32,  58,  62, 80, 67,  26,  44,
+	-5,  19,  26,  36, 17, 45,  61,  16,
+	-24, -11,   7,  26, 24, 35,  -8, -20,
+	-36, -26, -12,  -1,  9, -7,   6, -23,
+	-45, -25, -16, -17,  3,  0,  -5, -33,
+	-44, -16, -20,  -9, -1, 11,  -6, -71,
+	-19, -13,   1,  17, 16,  7, -37, -26,
+
+	// queen
+	-28,   0,  29,  12,  59,  44,  43,  45,
+	-24, -39,  -5,   1, -16,  57,  28,  54,
+	-13, -17,   7,   8,  29,  56,  47,  57,
+	-27, -27, -16, -16,  -1,  17,  -2,   1,
+	-9, -26,  -9, -10,  -2,  -4,   3,  -3,
+	-14,   2, -11,  -2,  -5,   2,  14,   5,
+	-35,  -8,  11,   2,   8,  15,  -3,   1,
+	-1, -18,  -9,  10, -15, -25, -31, -50,
+
+	// king
+	-65,  23,  16, -15, -56, -34,   2,  13,
+	29,  -1, -20,  -7,  -8,  -4, -38, -29,
+	-9,  24,   2, -16, -20,   6,  22, -22,
+	-17, -20, -12, -27, -30, -25, -14, -36,
+	-49,  -1, -27, -39, -46, -44, -33, -51,
+	-14, -14, -22, -46, -44, -30, -15, -27,
+	1,   7,  -8, -64, -43, -16,   9,   8,
+	-15,  36,  12, -54,   8, -28,  24,  14,
+
+
+	// Endgame positional piece scores //
+
+	//pawn
+	0,   0,   0,   0,   0,   0,   0,   0,
+	178, 173, 158, 134, 147, 132, 165, 187,
+	94, 100,  85,  67,  56,  53,  82,  84,
+	32,  24,  13,   5,  -2,   4,  17,  17,
+	13,   9,  -3,  -7,  -7,  -8,   3,  -1,
+	4,   7,  -6,   1,   0,  -5,  -1,  -8,
+	13,   8,   8,  10,  13,   0,   2,  -7,
+	0,   0,   0,   0,   0,   0,   0,   0,
+
+	// knight
+	-58, -38, -13, -28, -31, -27, -63, -99,
+	-25,  -8, -25,  -2,  -9, -25, -24, -52,
+	-24, -20,  10,   9,  -1,  -9, -19, -41,
+	-17,   3,  22,  22,  22,  11,   8, -18,
+	-18,  -6,  16,  25,  16,  17,   4, -18,
+	-23,  -3,  -1,  15,  10,  -3, -20, -22,
+	-42, -20, -10,  -5,  -2, -20, -23, -44,
+	-29, -51, -23, -15, -22, -18, -50, -64,
+
+	// bishop
+	-14, -21, -11,  -8, -7,  -9, -17, -24,
+	-8,  -4,   7, -12, -3, -13,  -4, -14,
+	2,  -8,   0,  -1, -2,   6,   0,   4,
+	-3,   9,  12,   9, 14,  10,   3,   2,
+	-6,   3,  13,  19,  7,  10,  -3,  -9,
+	-12,  -3,   8,  10, 13,   3,  -7, -15,
+	-14, -18,  -7,  -1,  4,  -9, -15, -27,
+	-23,  -9, -23,  -5, -9, -16,  -5, -17,
+
+	// rook
+	13, 10, 18, 15, 12,  12,   8,   5,
+	11, 13, 13, 11, -3,   3,   8,   3,
+	7,  7,  7,  5,  4,  -3,  -5,  -3,
+	4,  3, 13,  1,  2,   1,  -1,   2,
+	3,  5,  8,  4, -5,  -6,  -8, -11,
+	-4,  0, -5, -1, -7, -12,  -8, -16,
+	-6, -6,  0,  2, -9,  -9, -11,  -3,
+	-9,  2,  3, -1, -5, -13,   4, -20,
+
+	// queen
+	-9,  22,  22,  27,  27,  19,  10,  20,
+	-17,  20,  32,  41,  58,  25,  30,   0,
+	-20,   6,   9,  49,  47,  35,  19,   9,
+	3,  22,  24,  45,  57,  40,  57,  36,
+	-18,  28,  19,  47,  31,  34,  39,  23,
+	-16, -27,  15,   6,   9,  17,  10,   5,
+	-22, -23, -30, -16, -16, -23, -36, -32,
+	-33, -28, -22, -43,  -5, -32, -20, -41,
+
+	// king
+	-74, -35, -18, -18, -11,  15,   4, -17,
+	-12,  17,  14,  17,  17,  38,  23,  11,
+	10,  17,  23,  15,  20,  45,  44,  13,
+	-8,  22,  24,  27,  26,  33,  26,   3,
+	-18,  -4,  21,  24,  27,  23,   9, -11,
+	-19,  -3,  11,  21,  23,  16,   7,  -9,
+	-27, -11,   4,  13,  14,   4,  -5, -17,
+	-53, -34, -21, -11, -28, -14, -24, -43
+};
+
+// mirror positional score tables for opposite side only flip vertical
+constexpr int mirror_square[SQUARE_NB] =
+{
+	a1, b1, c1, d1, e1, f1, g1, h1,
+	a2, b2, c2, d2, e2, f2, g2, h2,
+	a3, b3, c3, d3, e3, f3, g3, h3,
+	a4, b4, c4, d4, e4, f4, g4, h4,
+	a5, b5, c5, d5, e5, f5, g5, h5,
+	a6, b6, c6, d6, e6, f6, g6, h6,
+	a7, b7, c7, d7, e7, f7, g7, h7,
+	a8, b8, c8, d8, e8, f8, g8, h8
+};
+
+//// mirror positional score tables for opposite side (flips vertical and horizontal)
+//const int mirror_square[SQUARE_NB] =
+//{
+//	h1, g1, f1, e1, d1, c1, b1, a1,
+//	h2, g2, f2, e2, d2, c2, b2, a2,
+//	h3, g3, f3, e3, d3, c3, b3, a3,
+//	h4, g4, f4, e4, d4, c4, b4, a4,
+//	h5, g5, f5, e5, d5, c5, b5, a5,
+//	h6, g6, f6, e6, d6, c6, b6, a6,
+//	h7, g7, f7, e7, d7, c7, b7, a7,
+//	h8, g8, f8, e8, d8, c8, b8, a8
+//};
+
+
+/*
+		  Rank mask            File mask           Isolated mask        Passed pawn mask
+		for square a6        for square f2         for square g2          for square c4
+	8  0 0 0 0 0 0 0 0    8  0 0 0 0 0 1 0 0    8  0 0 0 0 0 1 0 1     8  0 1 1 1 0 0 0 0
+	7  0 0 0 0 0 0 0 0    7  0 0 0 0 0 1 0 0    7  0 0 0 0 0 1 0 1     7  0 1 1 1 0 0 0 0
+	6  1 1 1 1 1 1 1 1    6  0 0 0 0 0 1 0 0    6  0 0 0 0 0 1 0 1     6  0 1 1 1 0 0 0 0
+	5  0 0 0 0 0 0 0 0    5  0 0 0 0 0 1 0 0    5  0 0 0 0 0 1 0 1     5  0 1 1 1 0 0 0 0
+	4  0 0 0 0 0 0 0 0    4  0 0 0 0 0 1 0 0    4  0 0 0 0 0 1 0 1     4  0 0 0 0 0 0 0 0
+	3  0 0 0 0 0 0 0 0    3  0 0 0 0 0 1 0 0    3  0 0 0 0 0 1 0 1     3  0 0 0 0 0 0 0 0
+	2  0 0 0 0 0 0 0 0    2  0 0 0 0 0 1 0 0    2  0 0 0 0 0 1 0 1     2  0 0 0 0 0 0 0 0
+	1  0 0 0 0 0 0 0 0    1  0 0 0 0 0 1 0 0    1  0 0 0 0 0 1 0 1     1  0 0 0 0 0 0 0 0
+	   a b c d e f g h       a b c d e f g h       a b c d e f g h        a b c d e f g h
+*/
+
+
+
+//// file masks [square]
+//U64 file_masks[SQUARE_NB];
+//
+//// rank masks [square]
+//U64 rank_masks[SQUARE_NB];
+//
+//// isolated pawn masks [square]
+//U64 isolated_masks[SQUARE_NB];
+//
+//// white passed pawn masks [square]
+//U64 white_passed_masks[SQUARE_NB];
+//
+//// black passed pawn masks [square]
+//U64 black_passed_masks[SQUARE_NB];
+
+
+
+
+// extract rank from a square [square]
+constexpr int get_rank[SQUARE_NB] =
+{
+	7, 7, 7, 7, 7, 7, 7, 7,
+	6, 6, 6, 6, 6, 6, 6, 6,
+	5, 5, 5, 5, 5, 5, 5, 5,
+	4, 4, 4, 4, 4, 4, 4, 4,
+	3, 3, 3, 3, 3, 3, 3, 3,
+	2, 2, 2, 2, 2, 2, 2, 2,
+	1, 1, 1, 1, 1, 1, 1, 1,
+	0, 0, 0, 0, 0, 0, 0, 0
+};
+
+// double pawns penalty
+constexpr int double_pawn_penalty_opening = -5;
+constexpr int double_pawn_penalty_endgame = -10;
+
+// isolated pawn penalty
+constexpr int isolated_pawn_penalty_opening = -5;
+constexpr int isolated_pawn_penalty_endgame = -10;
+
+// bonusses
+constexpr int passed_pawn_bonus[8] = { 0, 10, 30, 50, 75, 100, 150, 200 };
+
+// semi open file score
+constexpr int semi_open_file_score = 10;
+
+// open file score
+constexpr int open_file_score = 15;
+
+// mobility units (values from engine Fruit reloaded)
+constexpr const int bishop_unit = 4;
+constexpr const int queen_unit = 9;
+
+// mobility bonuses (values from engine Fruit reloaded)
+constexpr const int bishop_mobility_opening = 5;
+constexpr const int bishop_mobility_endgame = 5;
+constexpr const int queen_mobility_opening = 1;
+constexpr const int queen_mobility_endgame = 2;
+
+// king's shield bonus
+constexpr int king_shield_bonus = 5;
+
+
+
+int get_game_phase_score();
+int lerp_positional_score(int piece_type,int square,int game_phase_score);
+int lerp_material_score(int piece, int game_phase_score);
+int evaluate();
