+++ CrusherChess/src/search.h
@@ -1,0 +1,53 @@
+#pragma once
+
+#include "types.h"
+
+//interval for when to communicate with gui
+constexpr const int LISTEN = 16383;
+
+
+/*
+   most valuable victim & less valuable attacker
+
+	(Victims) Pawn Knight Bishop   Rook  Queen   King
+  (Attackers)
+		Pawn   105    205    305    405    505    605
+	  Knight   104    204    304    404    504    604
+	  Bishop   103    203    303    403    503    603
+		Rook   102    202    302    402    502    602
+	   Queen   101    201    301    401    501    601
+		King   100    200    300    400    500    600
+*/
+
+// MVV LVA [attacker][victim]
+constexpr const int mvv_lva[12][12] = {
+	105, 205, 305, 405, 505, 605,  105, 205, 305, 405, 505, 605,
+	104, 204, 304, 404, 504, 604,  104, 204, 304, 404, 504, 604,
+	103, 203, 303, 403, 503, 603,  103, 203, 303, 403, 503, 603,
+	102, 202, 302, 402, 502, 602,  102, 202, 302, 402, 502, 602,
+	101, 201, 301, 401, 501, 601,  101, 201, 301, 401, 501, 601,
+	100, 200, 300, 400, 500, 600,  100, 200, 300, 400, 500, 600,
+
+	105, 205, 305, 405, 505, 605,  105, 205, 305, 405, 505, 605,
+	104, 204, 304, 404, 504, 604,  104, 204, 304, 404, 504, 604,
+	103, 203, 303, 403, 503, 603,  103, 203, 303, 403, 503, 603,
+	102, 202, 302, 402, 502, 602,  102, 202, 302, 402, 502, 602,
+	101, 201, 301, 401, 501, 601,  101, 201, 301, 401, 501, 601,
+	100, 200, 300, 400, 500, 600,  100, 200, 300, 400, 500, 600
+};
+
+extern int killer_moves[2][MAX_PLY];
+extern int history_moves[12][SQUARE_NB];
+
+extern int pv_length[MAX_PLY];
+extern int pv_table[MAX_PLY][MAX_PLY];
+extern int follow_pv, score_pv;
+extern int move_scores[255];
+
+void search_position(int depth);
+int negamax(int alpha, int beta, int depth);
+int quiescence(int alpha, int beta);
+bool is_repetition();
+void sort_moves(Moves* move_list);
+int score_move(int move);
+void enable_pv_scoring(Moves* move_list);
